Some base Fragments code
Added some code for the fragments to work, though they're commented out
because there's not enough for them to be completed.
In this version opening a new page opens a new activity. Making this new
activity extend MainActivity and then copying all the OnCreate code
relating to the navigation drawer can get it to work on the new page.
Currently this version is stable and good to use in case of errors,
though will be heavily modified once the fragments are up and running.